py25/25-10-23/diary_app/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
PyMySQL==1.1.1
python-dotenv==1.0.1
Flask-Limiter==3.6.0
email-validator==2.2.0